Game over implementation
RoundEnd was displaying game over status, but was inside a check for no game over and wouldn't render. By showing round end when game is over, it was showing to start, which meant it was trying to run code on invalid data. Making another state for newGame makes code too complex and presents more problems, but moving GameOver to its own component is more logical. GameOver was heavily duplicate code from RoundEnd, and we still have to check if the game is valid in GameOver to not start new users with a game over screen (by checking bet.face === 1) so instead we just check wrapping the RoundEnd component. Checking if bet.face === 1 wrapping all components means that after starting a new game, the other states (pass the phone and player dashboard) don't render, so we must only check wrapping round end. Also needed to only call getNextPlayer if the game is not over, was infinite looping